daemon: Move *CredentialRequest/Response out of common
