+++ credentialsd/src/gateway/mod.rs
@@ -5,19 +5,17 @@ mod dbus;
 mod util;
 
 use std::{
+    collections::HashMap,
     path::{Path, PathBuf},
     sync::Arc,
 };
 
-use credentialsd_common::{
-    model::WebAuthnError,
-    server::{
-        CreateCredentialRequest, CreateCredentialResponse, GetCredentialRequest,
-        GetCredentialResponse, WindowHandle,
-    },
-};
+use credentialsd_common::{model::WebAuthnError, server::WindowHandle};
 use tokio::sync::Mutex as AsyncMutex;
-use zbus::Connection;
+use zbus::{
+    zvariant::{DeserializeDict, NoneValue, OwnedValue, SerializeDict, Type},
+    Connection,
+};
 
 use crate::{
     dbus::CredentialRequestController,
@@ -341,6 +339,94 @@ fn check_origin_from_privileged_client(
         _ => {
             tracing::warn!("Caller requested non-HTTPS schemed origin, which is not supported.");
             Err(WebAuthnError::SecurityError)
+        }
+    }
+}
+
+#[derive(Clone, Debug, DeserializeDict, Type)]
+#[zvariant(signature = "dict")]
+pub struct GetCredentialRequest {
+    pub origin: Option<String>,
+    pub is_same_origin: Option<bool>,
+    #[zvariant(rename = "publicKey")]
+    pub public_key: Option<GetPublicKeyCredentialRequest>,
+}
+
+#[derive(Clone, Debug, DeserializeDict, Type)]
+#[zvariant(signature = "dict")]
+pub struct GetPublicKeyCredentialRequest {
+    pub request_json: String,
+}
+
+#[derive(SerializeDict, Type)]
+#[zvariant(signature = "dict")]
+pub struct GetCredentialResponse {
+    #[zvariant(rename = "type")]
+    r#type: String,
+    public_key: Option<GetPublicKeyCredentialResponse>,
+}
+
+#[derive(SerializeDict, Type)]
+#[zvariant(signature = "dict")]
+pub struct GetPublicKeyCredentialResponse {
+    pub authentication_response_json: String,
+}
+
+impl From<GetPublicKeyCredentialResponse> for GetCredentialResponse {
+    fn from(response: GetPublicKeyCredentialResponse) -> Self {
+        GetCredentialResponse {
+            // TODO: Decide on camelCase or kebab-case for cred types
+            r#type: "public-key".to_string(),
+            public_key: Some(response),
+        }
+    }
+}
+
+#[derive(Clone, Debug, DeserializeDict, Type)]
+#[zvariant(signature = "dict")]
+pub struct CreateCredentialRequest {
+    pub origin: Option<String>,
+    pub is_same_origin: Option<bool>,
+    #[zvariant(rename = "type")]
+    pub r#type: String,
+    #[zvariant(rename = "publicKey")]
+    pub public_key: Option<CreatePublicKeyCredentialRequest>,
+}
+
+#[derive(SerializeDict, Type)]
+#[zvariant(signature = "dict")]
+pub struct CreateCredentialResponse {
+    #[zvariant(rename = "type")]
+    r#type: String,
+    public_key: Option<CreatePublicKeyCredentialResponse>,
+}
+
+impl NoneValue for CreateCredentialResponse {
+    type NoneType = HashMap<String, OwnedValue>;
+
+    fn null_value() -> Self::NoneType {
+        HashMap::new()
+    }
+}
+
+#[derive(Clone, Debug, DeserializeDict, Type)]
+#[zvariant(signature = "dict")]
+pub struct CreatePublicKeyCredentialRequest {
+    pub request_json: String,
+}
+
+#[derive(SerializeDict, Type)]
+#[zvariant(signature = "dict")]
+pub struct CreatePublicKeyCredentialResponse {
+    pub registration_response_json: String,
+}
+
+impl From<CreatePublicKeyCredentialResponse> for CreateCredentialResponse {
+    fn from(response: CreatePublicKeyCredentialResponse) -> Self {
+        CreateCredentialResponse {
+            // TODO: Decide on camelCase or kebab-case for cred types
+            r#type: "public-key".to_string(),
+            public_key: Some(response),
         }
     }
 }
